moved Component\BoostQuery to Component\DisMax\BoostQuery
